google_maps_ai_scraper/pyproject.toml
[project]
name = "pythonproject1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.78",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
]
